backend/employee-api/requirements.txt
psycopg[binary]==3.1.14
PyJWT==2.9.0